Add files via upload
Iteratively sets values of GOMAXPROCS, returns a MAP of GMP values to ping function runtimes. 
Pinger() and get_statistics() are left in there for now, but they should be replaced by Maria and Steve's code. 
An output channel called "out" is already created in my main(), but feel free to remove this in favor of another implementation.